Moved some class into the chart base class.
Also updated some of the in-code documentation.
